fix: abstract usePublicWallets
